data: Do not store calorie or macro counts in TrackedDay
There should be a single source of truth for all datda that persists
in the backend. Previously, the total calories and macro counts per
day were being stored twice - in the TrackedDay and Intake entities.
This introduces the potential for bugs that cause the TrackedDay
calories and macros to vary from the sum of Intake calories / macros.

Instead, remove tracked calorie and macro counts from TrackedDay
entities, and instead acquire these values by summing the Intake
entities.