bug: Fix bug with adding recent intake items again
When adding an item from the recent intakes list, the previously
added quantity and unit are pre-populated in the meal detail menu.
In the case where an item was added in units of serving sizes, the
quantity would be populated as the quantity in g or ml, but the unit
would be the full serving, resulting in the pre-populated nutrition
info not matching that of the previously added intake.