diary, cleanup: Use Visibility widget instead of conditional SizedBox
There are multiple ways to have conditional visibility for widgets.
One simple way is to use a ternary conditional to either populate the
widget, or an empty SizedBox. The other way is to use the Visibility
widget, which handles this internally (likely via the same
mechanism). Using a Visibility widget is clearer, and thus is the
preferred way to achieve conditional visibility.