bug: Fix async bug when adding a new tracked day
When logging food items on a given date, the marker that notifies the
user that there are tracked macro/calorie goals on that day (and
presumably also tracked intake items) was not appearing immediately.
This was because we were saving the trackedDay asynchronously, and
then triggering a LoadDiaryYearEvent before the new trackedDay was
saved.